# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Petar's Minecraft API",
    "description": "API for retrieving Minecraft item and block information (https://github.com/PetarMc1/Minecraft-Item-API)",
    "version": "1.0.0"
  },
  "basePath": "/",
  "paths": {
    "/api/items/{version}/{itemName}": {
      "get": {
        "summary": "Get item information by version and name",
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "description": "Minecraft version",
            "required": true,
            "type": "string",
            "enum": ["1.11", "1.12", "1.13", "1.14", "1.15", "1.16", "1.17", "1.18", "1.19", "1.20"]
          },
          {
            "name": "itemName",
            "in": "path",
            "description": "Name of the Minecraft item",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "schema": {
              "$ref": "#/definitions/Item"
            }
          },
          "404": {
            "description": "Item not found"
          }
        }
      }
    },
    "/api/blocks/{version}/{blockName}": {
      "get": {
        "summary": "Get block information by version and name",
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "description": "Minecraft version",
            "required": true,
            "type": "string",
            "enum": ["1.11", "1.12", "1.13", "1.14", "1.15", "1.16", "1.17", "1.18", "1.19", "1.20"]
          },
          {
            "name": "blockName",
            "in": "path",
            "description": "Name of the Minecraft block",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "schema": {
              "$ref": "#/definitions/Block"
            }
          },
          "404": {
            "description": "Block not found"
          }
        }
      }
    }
  },
  "definitions": {
    "Item": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        }
      }
    },
    "Block": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        }
      }
    }
  }
}